Remove hack in Friends Forever
An old hack in Friends Forever has finally been removed: the mod
used to try to detect the end of the day by relying on the player
going to sleep in the bed. This has been replaced by an event
that was added to SMAPI (at some point; I haven't been in the
modding scene for a while). That event can easily be used to
"pet" every pet and "talk" to every NPC.